requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
openai>=1.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0
